comment

+++ app/Http/Controllers/DashboardInspectorController.php
@@ -11585,7 +11585,12 @@ class DashboardInspectorController extends Controller
 		
 		if(!$exists)
 		{
-			$del = Users_location::where('user_id', auth()->user()->id)->whereNull('primary_owner')->delete();
+			/*$count = Task_list_corrective_action::where('lo_id', auth()->user()->id)->where('task_list_id', $task_list_id)->where('category_id', $category_id)->count();
+			
+			if($count==0)
+			{
+				$del = Users_location::where('user_id', auth()->user()->id)->whereNull('primary_owner')->delete();
+			}*/
 			
 			$userLocmodel = new Users_location();
 			$userLocmodel->user_id = $lo_id;
@@ -11595,18 +11600,78 @@ class DashboardInspectorController extends Controller
 			$userLocmodel->notification_status = 1;
 			$userLocmodel->save();
 		}
-		else{
-			$del = Users_location::where('user_id', auth()->user()->id)->whereNull('primary_owner')->delete();
-			Task_list_corrective_action::where('lo_id', auth()->user()->id)->where('task_list_id', $task_list_id)->where('category_id', $category_id)->delete();
+		else
+		{
+			/*$count = Task_list_corrective_action::where('lo_id', auth()->user()->id)->where('task_list_id', $task_list_id)->where('category_id', $category_id)->count();
+			
+			if($count==0)
+			{
+				$del = Users_location::where('user_id', auth()->user()->id)->whereNull('primary_owner')->delete();
+			}*/
+			
+			
+			// 07-10-2025
+			//Task_list_corrective_action::where('lo_id', auth()->user()->id)->where('task_list_id', $task_list_id)->where('category_id', $category_id)->delete();
 			//$del = false;
 		}
 		
-		/*Users_location::where('user_id', auth()->user()->id)->where('company_id', $company_id)->where('user_type', 2)->where('location_id', $location_id)->update(['user_id'=>$user_id, 'notification_status'=>1]);*/
+		$query = Task_list_corrective_action::where('lo_id', auth()->user()->id)->where('task_list_id', $task_list_id)->where('category_id', $category_id)->where('checklist_id', $checklist_id);
 		
+		if(!empty($subchecklist_id))
+		{
+			$query->where('subchecklist_id', $subchecklist_id);
+		}
 		
+		$isCorrectiveAction = $query->first();
 		
+		if($isCorrectiveAction)
+		{
+			$corrective_action_data = Task_list_corrective_action::where('task_list_id',$task_list_id )->where('category_id', $category_id)->where('checklist_id', $checklist_id)->where('subchecklist_id', $subchecklist_id)->where('lo_id', auth()->user()->id)->where('inspector_id', $inspector_id)->where('los_id', $los_id)->first();
+			
+			$corrective_action_id = $corrective_action_data ? $corrective_action_data->id : '';
+			
+			if($corrective_action_id)
+			{
+				$model = Task_list_corrective_action::find($corrective_action_id);
+				$model->task_list_id = $task_list_id;
+				$model->category_id = $category_id;
+				$model->checklist_id = $checklist_id;
+				$model->subchecklist_id = $subchecklist_id;
+				$model->lo_id = $lo_id ?? null;
+				$model->lo_corrective_action_plan = null;
+				$model->lo_completed_by = date('Y-m-d h:i:s');
+				$model->lo_direct_approve = 3;
+				$model->inspector_id = $inspector_id;
+				$model->los_id = $los_id;
+				$model->rejected_repeated = 1; // 04-08-2025
+				$model->tab_no = 1; // 04-08-2025
+				$model->created_at = date('Y-m-d h:i:s'); // 11-08-2025
+				$model->save();
+			}
+		}
+		else
+		{
+			$model = new Task_list_corrective_action();
+			$model->task_list_id = $task_list_id;
+			$model->category_id = $category_id;
+			$model->checklist_id = $checklist_id;
+			$model->subchecklist_id = $subchecklist_id;
+			$model->lo_id = $lo_id ?? null;
+			$model->lo_corrective_action_plan = null;
+			$model->lo_completed_by = date('Y-m-d h:i:s');
+			$model->lo_direct_approve = 3;
+			$model->inspector_id = $inspector_id;
+			$model->los_id = $los_id;
+			$model->rejected_repeated = 1; // 04-08-2025
+			$model->tab_no = 1; // 04-08-2025
+			$model->created_at = date('Y-m-d h:i:s'); // 11-08-2025
+			$model->save();
+			$id = $model->id;
+			
+		}
 		
-		if($del != false)
+		//----07-10-2025----
+		/*if($del != false)
 		{
 			$corrective_action_data = Task_list_corrective_action::where('task_list_id',$task_list_id )->where('category_id', $category_id)->where('checklist_id', $checklist_id)->where('subchecklist_id', $subchecklist_id)->where('lo_id', auth()->user()->id)->where('inspector_id', $inspector_id)->where('los_id', $los_id)->first();
 			
@@ -11652,6 +11717,13 @@ class DashboardInspectorController extends Controller
 			$model->created_at = date('Y-m-d h:i:s'); // 11-08-2025
 			$model->save();
 			$id = $model->id;
+		}*/
+		
+		$count = Task_list_corrective_action::where('lo_id', auth()->user()->id)->where('task_list_id', $task_list_id)->where('category_id', $category_id)->count();
+			
+		if($count==0)
+		{
+			$del = Users_location::where('user_id', auth()->user()->id)->whereNull('primary_owner')->delete();
 		}
 		
 		/*$correctiveActionDtldModel = new Task_list_corrective_action_details();
